test: fix comparison between signed/unsigned integers
Lets avoid comparing signed and unsigned integers. Always convert to
ssize_t or ptrdiff_t.

+++ src/test-enumerated.c
@@ -50,7 +50,7 @@ static void test_enumerated_types(void) {
 
                 do {
                         c = dbus_typenum_step(e);
-                        if (pos - str >= n_str) {
+                        if (pos - str >= (ssize_t)n_str) {
                                 str = realloc(str, (pos - str) * 2);
                                 assert(str);
                                 pos = str + n_str;
